いろいろ切り分けたりなんだり
