+++ src/jsMain/kotlin/ktshsResearchAssignmentE2022/com/github/minesweeper/Main.kt
@@ -1,15 +1,13 @@
 package ktshsResearchAssignmentE2022.com.github.minesweeper
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import kotlinx.browser.window
 import ktshsResearchAssignmentE2022.com.github.minesweeper.components.SettingMenuButton
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.MineSweeperState
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.WindowState
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.SidebarState
 import ktshsResearchAssignmentE2022.com.github.minesweeper.styleSheets.*
 import org.jetbrains.compose.web.css.Style
 import org.jetbrains.compose.web.renderComposable
-
-var isSidebarOpen by mutableStateOf(false)
 
 fun main() {
     renderComposable("root") {
@@ -19,30 +17,28 @@ fun main() {
         Style(SidebarStyleSheet)
         Style(PhoneSidebarStyleSheet)
 
-        val outerWidth = mutableStateOf(window.outerWidth)
-        val visitorIsPhone = outerWidth.value < 850
-
         window.onresize = {
-            outerWidth.value = window.outerWidth
+            WindowState.width = window.outerWidth
+            WindowState.isPhone = WindowState.width < 850
             it
         }
 
-        if (MineSweeper.logic.isGameOver || MineSweeper.logic.isGameClear) ResultContainerLayout { Result() }
+        if (MineSweeperState.logic.isGameOver || MineSweeperState.logic.isGameClear) ResultContainerLayout { Result() }
 
-        if (visitorIsPhone) PhoneHamburgerButtonLayout {
-            if (isSidebarOpen) {
+        if (WindowState.isPhone) PhoneHamburgerButtonLayout {
+            if (SidebarState.isOpen) {
                 PhoneSidebar()
             } else {
                 SettingMenuButton {
-                    isSidebarOpen = true
+                    SidebarState.isOpen = true
                 }
             }
         }
 
         MainLayout {
-            if (!visitorIsPhone) PcSidebar()
+            if (!WindowState.isPhone) PcSidebar()
             CenterLayout {
-                MineSweeper.show()
+                MineSweeper()
             }
         }
     }
