+++ src/jsMain/kotlin/ktshsResearchAssignmentE2022/com/github/minesweeper/Settings.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import ktshsResearchAssignmentE2022.com.github.minesweeper.components.GrowingButton
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.MineSweeperState
 import ktshsResearchAssignmentE2022.com.github.minesweeper.styleSheets.SidebarStyleSheet
 import org.jetbrains.compose.web.attributes.InputType
 import org.jetbrains.compose.web.attributes.max
@@ -79,27 +80,27 @@ private fun SimpleSettings() {
 
         GrowingButton("かんたん", SettingState.difficulty == Difficulty.Easy) {
             SettingState.setWithDifficulty(Difficulty.Easy)
-            MineSweeper.regenerate()
+            MineSweeperState.regenerate()
         }
 
         GrowingButton("ふつう", SettingState.difficulty == Difficulty.Normal) {
             SettingState.setWithDifficulty(Difficulty.Normal)
-            MineSweeper.regenerate()
+            MineSweeperState.regenerate()
         }
 
         GrowingButton("むずかしい", SettingState.difficulty == Difficulty.Hard) {
             SettingState.setWithDifficulty(Difficulty.Hard)
-            MineSweeper.regenerate()
+            MineSweeperState.regenerate()
         }
     }
 }
 
 @Composable
 private fun AdvancedSettings() {
-    val notIsDiff = !(SettingState.yLength == MineSweeper.logic.yLength &&
-            SettingState.xLength == MineSweeper.logic.xLength &&
-            SettingState.numOfMines == MineSweeper.logic.numOfMines &&
-            SettingState.seed == MineSweeper.logic.seed)
+    val notIsDiff = !(SettingState.yLength == MineSweeperState.logic.yLength &&
+            SettingState.xLength == MineSweeperState.logic.xLength &&
+            SettingState.numOfMines == MineSweeperState.logic.numOfMines &&
+            SettingState.seed == MineSweeperState.logic.seed)
 
     Div({
         style {
@@ -115,15 +116,15 @@ private fun AdvancedSettings() {
             Text("詳細設定")
         }
         GrowingButton("再生成する", notIsDiff) {
-            if (notIsDiff) MineSweeper.regenerate()
+            if (notIsDiff) MineSweeperState.regenerate()
         }
         GrowingButton("現在の設定にもどす", notIsDiff, 10.percent) {
             if (notIsDiff) {
                 SettingState.setAll(
-                    MineSweeper.logic.yLength,
-                    MineSweeper.logic.xLength,
-                    MineSweeper.logic.numOfMines,
-                    MineSweeper.logic.seed
+                    MineSweeperState.logic.yLength,
+                    MineSweeperState.logic.xLength,
+                    MineSweeperState.logic.numOfMines,
+                    MineSweeperState.logic.seed
                 )
             }
         }
