+++ src/jsMain/kotlin/ktshsResearchAssignmentE2022/com/github/minesweeper/Sidebar.kt
@@ -2,10 +2,12 @@ package ktshsResearchAssignmentE2022.com.github.minesweeper
 
 import androidx.compose.runtime.Composable
 import ktshsResearchAssignmentE2022.com.github.minesweeper.components.GrowingButton
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.MineSweeperState
 import ktshsResearchAssignmentE2022.com.github.minesweeper.styleSheets.SidebarStyleSheet
 import org.jetbrains.compose.web.css.percent
 import org.jetbrains.compose.web.css.textAlign
 import org.jetbrains.compose.web.dom.*
+import ktshsResearchAssignmentE2022.com.github.minesweeper.states.SidebarState
 
 @Composable
 fun PcSidebar() {
@@ -15,11 +17,11 @@ fun PcSidebar() {
         Style({
             classes(SidebarStyleSheet.elementStyle)
         }) {
-            H1 { Text(if (!MineSweeper.logic.isDevMode) "まいんすいーぱー" else "Dev Mode") }
+            H1 { Text(if (!MineSweeperState.logic.isDevMode) "まいんすいーぱー" else "Dev Mode") }
             Settings()
             Button({
                 onClick {
-                    MineSweeper.logic.isDevMode = !MineSweeper.logic.isDevMode
+                    MineSweeperState.logic.isDevMode = !MineSweeperState.logic.isDevMode
                 }
             }) {
                 H3({
@@ -44,14 +46,14 @@ fun PhoneSidebar() {
                 style {
                     textAlign("center")
                 }
-            }) { Text(if (!MineSweeper.logic.isDevMode) "まいんすいーぱー" else "Dev Mode") }
+            }) { Text(if (!MineSweeperState.logic.isDevMode) "まいんすいーぱー" else "Dev Mode") }
             Settings()
             GrowingButton("閉じる", true, 10.percent) {
-                isSidebarOpen = false
+                SidebarState.isOpen = false
             }
             Button({
                 onClick {
-                    MineSweeper.logic.isDevMode = !MineSweeper.logic.isDevMode
+                    MineSweeperState.logic.isDevMode = !MineSweeperState.logic.isDevMode
                 }
             }) {
                 H3 { Text("Made by 神奈川工業高校電気科") }
