+++ src/jsMain/kotlin/ktshsResearchAssignmentE2022/com/github/minesweeper/states/SidebarState.kt
@@ -1,0 +1,9 @@
+package ktshsResearchAssignmentE2022.com.github.minesweeper.states
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+
+object SidebarState {
+    var isOpen by mutableStateOf(false)
+}
